feat: changes applied:      * improve js code to intercept ruby create method      * change WebAuthnJSON import for a more complete one      * change js controller name
