+++ app/controllers/users/passkeys_controller.rb
@@ -1,10 +1,7 @@
 module Users
   class PasskeysController < ApplicationController
     def index
-    end
-
-    def create
-      create_options = WebAuthn::Credential.options_for_create(
+      @create_passkey_options = WebAuthn::Credential.options_for_create(
         user: {
           id: current_user.webauthn_id,
           name: current_user.email.split('@').first
@@ -12,13 +9,11 @@ module Users
         exclude: current_user.passkeys.pluck(:external_id),
         authenticator_selection: { user_verification: "required" }
       )
-      session[:current_registration_challenge] = { challenge: create_options.challenge }
-
-      render json: create_options
+      session[:current_registration_challenge] = { challenge: @create_passkey_options.challenge }
     end
 
-    def callback
-      webauthn_passkey = WebAuthn::Credential.from_create(params)
+    def create
+      webauthn_passkey = WebAuthn::Credential.from_create(JSON.parse(params[:passkey_public_key]))
 
       begin
         webauthn_passkey.verify(session[:current_registration_challenge]["challenge"], user_verification: true)
@@ -28,7 +23,7 @@ module Users
         )
 
         if passkey.update(
-          name: params[:credential_name],
+          name: params[:name],
           public_key: webauthn_passkey.public_key,
           sign_count: webauthn_passkey.sign_count
         )
